+++ src/modules/public-pages/index.js
@@ -1,0 +1,1 @@
+export { PublicPages } from './PublicPages.js';