+++ src/modules/seo/index.js
@@ -1,0 +1,1 @@
+export { SEOManager } from './SEOManager.js';